Update
I'm currently building a UI input condition for the bulk node method, and am trying to get user input on how many nodes they wish to create and then inittialize and array of that size. with that we can then iterate through the list and input data into each node.

the problem is that i can't just put a variable into the list size, it has to be const. so what im thinking i do is that i simply create a function that converts vars to const (or mimicks such).